Aluno.java
Simple java exercise i did a while ago, and only publishing now. Although simple, i think this code got a good logic and i think i did a good work with it. The code's main goal is to print a student's informations, like his name, his registry number and his grades. The only "complex" part of the code is the method of making the averages of the student's grades. It gets the student's grades for the 1st and 2nd test he did and calculates the averages individually for each discipline.